requirements.txt
discord.py>=2.3.2
asyncpg>=0.29.0
python-dotenv>=1.0.1
